refactor: 채팅방과 채팅 메시지 페이지들이 layout 공유가 되지 않으므로 불필요한 layout 파일 제거 (#191)
