+++ apps/web/app/(pages)/chat/[chatId]/page.tsx
@@ -1,3 +1,10 @@
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: '채팅방 - 경매 플랫폼',
+  description: '실시간 채팅으로 거래를 진행하세요.',
+};
+
 // 채팅 방 페이지
 const ChatRoomPage = ({ params }: { params: { chatId: string } }) => {
   return (
