+++ apps/web/app/(pages)/chat/page.tsx
@@ -1,6 +1,29 @@
-// 채팅 리스트 페이지
+import type { Metadata } from 'next';
+
+import { TopNavigation, BottomNavigation, PageContainer } from '@/components/layout';
+
+export const metadata: Metadata = {
+  title: '채팅 목록 - 경매 플랫폼',
+  description: '진행 중인 채팅 목록을 확인하세요.',
+};
+
 const ChatListPage = () => {
-  return <div>준비중입니다!</div>;
+  return (
+    <div className="h-screen flex flex-col">
+      <TopNavigation
+        title="채팅"
+        showTitle
+        showAlarmButton
+        showBackButton={false}
+        showSearchButton={false}
+        showRegion={false}
+      />
+      <PageContainer className="py-lg">
+        <div></div>
+      </PageContainer>
+      <BottomNavigation />
+    </div>
+  );
 };
 
 export default ChatListPage;
